Fixed privileges problem, you cannot setup redirect if you are not root
