+++ src/ec_sslwrap.c
@@ -29,6 +29,7 @@
 #include <ec_file.h>
 #include <ec_version.h>
 #include <ec_socket.h>
+#include <ec_utils.h>
 
 #include <sys/types.h>
 #ifndef OS_WINDOWS
@@ -391,7 +392,9 @@ static int sslw_insert_redirect(u_int16 sport, u_int16 dport)
    /* execute the script */ 
    switch (fork()) {
       case 0:
+         regain_privs();
          execvp(param[0], param);
+         drop_privs();
          WARN_MSG("Cannot setup http redirect (command: %s), please edit your etter.conf file and put a valid value in redir_command_on field\n", param[0]);
          safe_free_mem(param, &param_length, command);
          _exit(EINVALID);
@@ -459,7 +462,9 @@ static int sslw_remove_redirect(u_int16 sport, u_int16 dport)
    /* execute the script */ 
    switch (fork()) {
       case 0:
+         regain_privs();
          execvp(param[0], param);
+         drop_privs();
          WARN_MSG("Cannot remove http redirect (command: %s), please edit your etter.conf file and put a valid value in redir_command_on field\n", param[0]);
          safe_free_mem(param, &param_length, command);
          _exit(EINVALID);
